account for suggestions

+++ src/Vehicle/FTPManager.cc
@@ -630,7 +630,7 @@ void FTPManager::_sendRequestExpectAck(MavlinkFTP::Request* request)
 bool FTPManager::_parseURI(uint8_t fromCompId, const QString& uri, QString& parsedURI, uint8_t& compId)
 {
     parsedURI   = uri;
-    compId      = (fromCompId == MAV_COMP_ID_ALL) ? MAV_COMP_ID_AUTOPILOT1 : fromCompId;
+    compId      = (fromCompId == MAV_COMP_ID_ALL) ? (uint8_t)MAV_COMP_ID_AUTOPILOT1 : fromCompId;
 
     // Pull scheme off the front if there
     QString ftpPrefix(QStringLiteral("%1://").arg(mavlinkFTPScheme));
